Share DataGrid column-filter matcher across Lite/Dashboard
Fourth wave of the code-sharing follow-up. The operator-based ColumnFilterState
matcher (MatchesFilter + IsValueEmpty/CompareValues/CompareNumeric/
TryParseNumeric) was byte-identical in Lite's and Dashboard's DataGridFilterService.
Extracted it to PerformanceMonitor.Common.ColumnFilterMatcher (pure reflection +
value comparison, no WPF dependency).

Both apps' DataGridFilterService.MatchesFilter(item, filter) is now a one-line
forwarder to the shared matcher, so the ~28 Dashboard + Lite call sites are
untouched. Dashboard keeps its app-specific extras (the string-overload
MatchesFilter, ApplyFilter, NumericTypes/DateTypes, property cache); Lite's
service collapses to the single forwarder.

DataGridFilterManager is intentionally NOT touched here — it carries a behavioral
diff (Lite preserves sort on filter, Dashboard doesn't), which is a reconciliation
decision for a separate change.

Pure refactor, no behavior change. Solution builds clean; Dashboard.Tests (488),
Lite.Tests (544), Installer.Tests (80) all pass.
